requirements.txt
matplotlib==3.4.3
numpy==1.22.3
onnx==1.11.0
onnxruntime==1.14.1
onnxruntime_gpu==1.12.1
pandas==1.4.1
Pillow==9.5.0
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.11.2
tensorflow==2.8.0
tvm==1.0.0
